API側モデルの分割
