+++ backend/app/api/api_v1/routers/topics.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, Request, Response
 
 from app.core.auth import get_current_active_user
 from app.db.crud import create_topic, drop_topic, edit_topic, get_topics
-from app.db.schemas import Topic, TopicCreate, TopicEdit
+from app.db.schemas.topics import Topic, TopicCreate, TopicEdit
 from app.db.session import get_db
 
 topics_router = r = APIRouter()
