+++ backend/app/initial_data.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 from app.db.crud import create_user
-from app.db.schemas import UserCreate
+from app.db.schemas.users import UserCreate
 from app.db.session import SessionLocal
